pyproject.toml
[project]
name = "mcp-reminders"
version = "0.1.0"
description = "A Model Context Protocol server providing tools for CRUD operations for Apple Reminders"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "mcp[cli]>=1.2.1",
    "pyobjc-core>=11.0",
    "pyobjc-framework-EventKit>=11.0",
    "pyobjc-framework-Cocoa>=11.0"
]

[dependency-groups]
dev = ["pytest>=8.3.4", "pytest-mock>=3.14.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"
